Allow multiple operations with a single connection
Performance Improvement: Allow a single PowerShell to be sent multiple
commands before it's closed, saving the overhead of starting a new one
for every command. (For simplicity, apply the same API change to the
VBSEngine provider even though it has less overhead and so has not yet
been restructured.)

For compatibility (and to provide a simpler API in single-operation
cases), provide a set of delegate methods that build a connection,
perform one operation, then close it (as before).

Change Java language level to 8 (from 5) to use AutoCloseable.

Also, fix some related grammatical errors and typos in comments.